Move to non conj

+++ irm/group_coords.h
@@ -6,7 +6,7 @@
 
 namespace irm { 
 
-typedef int groupid_t; 
+typedef int16_t groupid_t; 
 static const int MAX_AXES = 4; 
 
 #ifdef GROUP_COORDS_ARRAY
